requirements.txt
PyQt6~=6.2.2
requests~=2.26.0
appscript~=1.1.2
PyYAML~=6.0
modulegraph~=0.19.2
py2app~=0.26.1
macholib~=1.15.2
setuptools~=60.2.0
altgraph~=0.17.2
certifi~=2021.10.8
APScheduler~=3.8.1
pytz~=2021.3